requirements.txt
# Legal Mediation System - Consolidated Dependencies
# Install with: pip install -r requirements.txt
# Python 3.10+ required

# ============================================
# Core Dependencies
# ============================================

# Data validation and configuration
pydantic>=2.5.0
python-dotenv>=1.0.0

# Logging
structlog>=23.2.0

# CLI
click>=8.1.0

# Async utilities
aiofiles>=23.2.0
numpy>=1.24.0

# ============================================
# RAG Engine Dependencies
# ============================================

# PDF extraction
pymupdf>=1.24.0

# Embeddings (OpenAI)
openai>=1.0.0
tiktoken>=0.5.0

# Vector store
chromadb>=0.4.0

# Keyword search
rank-bm25>=0.2.2

# ============================================
# LLM Orchestrator Dependencies
# ============================================

# Anthropic Claude API
anthropic>=0.39.0

# ============================================
# Knowledge Graph Builder Dependencies
# ============================================

# Note: Neo4j will be added when we migrate from JSON storage
# neo4j>=5.0.0

# ============================================
# API Layer Dependencies
# ============================================

# FastAPI
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# Supabase (storage and auth)
supabase>=2.0.0

# ============================================
# Web Scraper Dependencies (scripts/)
# ============================================

# Async HTTP client
aiohttp>=3.9.0

# HTML parsing
beautifulsoup4>=4.12.0
lxml>=5.0.0  # Faster HTML parser backend

# Retry logic with exponential backoff
tenacity>=8.2.0

# Progress bars
tqdm>=4.66.0

# Rate limiting
aiolimiter>=1.1.0

# SQLite async support
aiosqlite>=0.19.0

# For robots.txt parsing
robotexclusionrulesparser>=1.7.0

# ============================================
# Testing Dependencies
# ============================================
pytest>=8.0.0
pytest-asyncio>=0.23.0

# ============================================
# Development Tools (Optional)
# ============================================
# Uncomment for development
# black>=23.0.0
# ruff>=0.1.0
# mypy>=1.0.0
